Container does now make use of withRedux
Before `Container` component was exported by default wrapped with helper
function `withRedux()`. It was not the right way to handle it so as I
was getting a `WrappedComponent` from `connect()` to work with and that
had already been initiated.

Now we do some inception work instead using our own `Container`
component to wrap ourselves and return a wrapped version of it with
`withRedux()`.